utils/ml_model.py
import os
import numpy as np
from PIL import Image
import cv2
import random

# TensorFlow imports commented out for demo - uncomment when TensorFlow is installed
# import tensorflow as tf
# from tensorflow.keras.applications import DenseNet121
# from tensorflow.keras.preprocessing import image
# from tensorflow.keras.applications.densenet import preprocess_input

class MLPredictor:
    def __init__(self):
        self.models = {}
        self.class_labels = {
            'chest': ['Normal', 'Pneumonia', 'COVID-19', 'Tuberculosis', 'Lung Cancer'],
            'hand': ['Normal', 'Fracture', 'Arthritis', 'Dislocation'],
            'leg': ['Normal', 'Fracture', 'Arthritis', 'Bone Tumor'],
            'skull': ['Normal', 'Fracture', 'Tumor', 'Hemorrhage'],
            'spine': ['Normal', 'Fracture', 'Scoliosis', 'Disc Herniation'],
            'pelvis': ['Normal', 'Fracture', 'Hip Dysplasia', 'Arthritis']
        }
        self.load_models()
    
    def load_models(self):
        """Load pre-trained models for different body parts"""
        # For demo purposes, we'll use dummy models
        # In production, you would load actual TensorFlow/PyTorch models
        try:
            # TensorFlow model loading code (commented out for demo)
            # base_model = DenseNet121(weights='imagenet', include_top=False, input_shape=(224, 224, 3))
            # model = tf.keras.Sequential([...])

            # For demo, create dummy models
            self.create_dummy_models()
            print("Demo ML models loaded successfully!")

        except Exception as e:
            print(f"Error loading ML models: {e}")
            # Create dummy models for development
            self.create_dummy_models()
    
    def create_dummy_models(self):
        """Create dummy models for development/testing"""
        print("Creating dummy models for development...")
        for body_part in self.class_labels.keys():
            self.models[body_part] = None
    
    def preprocess_image(self, image_path):
        """Preprocess image for model prediction"""
        try:
            # Load and resize image
            img = Image.open(image_path).convert('RGB')
            img = img.resize((224, 224))

            # Convert to array (TensorFlow preprocessing commented out for demo)
            img_array = np.array(img)
            img_array = np.expand_dims(img_array, axis=0)
            # img_array = preprocess_input(img_array)  # Uncomment when using TensorFlow

            return img_array

        except Exception as e:
            print(f"Error preprocessing image: {e}")
            return None
    
    def predict(self, image_path, body_part):
        """Make prediction on X-ray image"""
        try:
            # Preprocess image
            processed_image = self.preprocess_image(image_path)
            if processed_image is None:
                return None, 0.0
            
            # Get model for body part
            model = self.models.get(body_part)
            if model is None:
                # Return dummy prediction for development
                return self.dummy_prediction(body_part)
            
            # Make prediction
            predictions = model.predict(processed_image)
            predicted_class_idx = np.argmax(predictions[0])
            confidence = float(predictions[0][predicted_class_idx])
            
            # Get class label
            class_labels = self.class_labels[body_part]
            predicted_class = class_labels[predicted_class_idx]
            
            return predicted_class, confidence
            
        except Exception as e:
            print(f"Error making prediction: {e}")
            return self.dummy_prediction(body_part)
    
    def dummy_prediction(self, body_part):
        """Return dummy prediction for development"""
        import random
        
        class_labels = self.class_labels[body_part]
        predicted_class = random.choice(class_labels)
        confidence = random.uniform(0.7, 0.95)
        
        return predicted_class, confidence
    
    def validate_body_part_image(self, image_path, expected_body_part):
        """Validate if image matches expected body part"""
        # This is a simplified validation
        # In production, you would use a separate model to classify body parts
        return True
    
    def get_supported_body_parts(self):
        """Get list of supported body parts"""
        return list(self.class_labels.keys())
    
    def get_class_labels(self, body_part):
        """Get class labels for a specific body part"""
        return self.class_labels.get(body_part, [])
