requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
Pillow==10.0.1
tensorflow>=2.16.0
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
googletrans==4.0.0rc1
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
